move TransactionInfo to starcoin-types
