+++ test-helper/tests/types_check.rs
@@ -9,14 +9,16 @@ use serde::Serialize;
 use starcoin_crypto::hash::PlainCryptoHash;
 use starcoin_crypto::HashValue;
 use starcoin_logger::prelude::*;
-use starcoin_types::block::{Block, BlockHeader, BlockInfo};
-use starcoin_types::compact_block::CompactBlock;
-use starcoin_types::startup_info::ChainStatus;
-use starcoin_vm_types::block_metadata::BlockMetadata;
-use starcoin_vm_types::transaction::{
-    Module, Package, RawUserTransaction, Script, ScriptFunction, SignedUserTransaction,
-    TransactionInfo,
+use starcoin_types::{
+    block::{Block, BlockHeader, BlockInfo},
+    compact_block::CompactBlock,
+    startup_info::ChainStatus,
+    transaction::{
+        Module, Package, RawUserTransaction, Script, ScriptFunction, SignedUserTransaction,
+        TransactionInfo,
+    },
 };
+use starcoin_vm_types::block_metadata::BlockMetadata;
 use std::any::type_name;
 use std::path::PathBuf;
