+++ vm/types/src/transaction/mod.rs
@@ -18,7 +18,6 @@ use crate::{
 use anyhow::{format_err, Error, Result};
 use bcs_ext::Sample;
 use serde::{Deserialize, Deserializer, Serialize};
-use starcoin_accumulator::inmemory::InMemoryAccumulator;
 use starcoin_crypto::multi_ed25519::{MultiEd25519PublicKey, MultiEd25519Signature};
 use starcoin_crypto::{
     ed25519::*,
@@ -42,7 +41,6 @@ pub use script::{
     ArgumentABI, Script, ScriptABI, ScriptFunction, ScriptFunctionABI, TransactionScriptABI,
     TypeArgumentABI,
 };
-use starcoin_crypto::hash::SPARSE_MERKLE_PLACEHOLDER_HASH;
 use std::str::FromStr;
 pub use transaction_argument::{
     parse_transaction_argument, parse_transaction_arguments, TransactionArgument,
@@ -728,94 +726,6 @@ impl TransactionOutput {
             }
         }
         table_items
-    }
-}
-
-/// `TransactionInfo` is the object we store in the transaction accumulator. It consists of the
-/// transaction as well as the execution result of this transaction.
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, CryptoHasher, CryptoHash)]
-pub struct TransactionInfo {
-    /// The hash of this transaction.
-    pub transaction_hash: HashValue,
-
-    /// The root hash of Sparse Merkle Tree describing the world state at the end of this
-    /// transaction.
-    pub state_root_hash: HashValue,
-
-    /// The root hash of Merkle Accumulator storing all events emitted during this transaction.
-    pub event_root_hash: HashValue,
-
-    /// The amount of gas used.
-    pub gas_used: u64,
-
-    /// The vm status. If it is not `Executed`, this will provide the general error class. Execution
-    /// failures and Move abort's receive more detailed information. But other errors are generally
-    /// categorized with no status code or other information
-    pub status: KeptVMStatus,
-}
-
-impl TransactionInfo {
-    /// Constructs a new `TransactionInfo` object using transaction hash, state root hash and event
-    /// root hash.
-    pub fn new(
-        transaction_hash: HashValue,
-        state_root_hash: HashValue,
-        events: &[ContractEvent],
-        gas_used: u64,
-        status: KeptVMStatus,
-    ) -> TransactionInfo {
-        let event_hashes: Vec<_> = events.iter().map(|e| e.crypto_hash()).collect();
-        let events_accumulator_hash =
-            InMemoryAccumulator::from_leaves(event_hashes.as_slice()).root_hash();
-        TransactionInfo {
-            transaction_hash,
-            state_root_hash,
-            event_root_hash: events_accumulator_hash,
-            gas_used,
-            status,
-        }
-    }
-
-    pub fn id(&self) -> HashValue {
-        self.crypto_hash()
-    }
-
-    /// Returns the hash of this transaction.
-    pub fn transaction_hash(&self) -> HashValue {
-        self.transaction_hash
-    }
-
-    /// Returns root hash of Sparse Merkle Tree describing the world state at the end of this
-    /// transaction.
-    pub fn state_root_hash(&self) -> HashValue {
-        self.state_root_hash
-    }
-
-    /// Returns the root hash of Merkle Accumulator storing all events emitted during this
-    /// transaction.
-    pub fn event_root_hash(&self) -> HashValue {
-        self.event_root_hash
-    }
-
-    /// Returns the amount of gas used by this transaction.
-    pub fn gas_used(&self) -> u64 {
-        self.gas_used
-    }
-
-    pub fn status(&self) -> &KeptVMStatus {
-        &self.status
-    }
-}
-
-impl Sample for TransactionInfo {
-    fn sample() -> Self {
-        Self::new(
-            SignedUserTransaction::sample().id(),
-            *SPARSE_MERKLE_PLACEHOLDER_HASH,
-            &[],
-            0,
-            KeptVMStatus::Executed,
-        )
     }
 }
